Moved reusable code for CS+ into Base.h and Base.cpp
